services/news_feed/requirements.txt
grpcio
crewai
langchain_community
ddgs
